convert float arrays to vectors, remove old util functions that took float arrays, change pinball shape origin to center so that i dont have to do our own vector rotation. this commit has no functionality but its an insane amount of cleanup

+++ games/pinball/source/pinball.h
@@ -21,7 +21,7 @@ class Pinball {
         // Box2D works best with objects between 0.1m and 10m in size.
         // A pinball is normally 0.027 meters in diameter, but this is too small for Box2D.
         // We simulate a pinball with a diameter of 0.16m and scale the pinball machine around this.
-        double m_radius = 0.08;
+        float m_radius = 0.08;
 
         // Coordinates for the launch tube.
         float m_startX = 2.0f;
